refactor(datasource): remove legacy gdrive skill
Replace the duplicate Google Drive REST and rclone compatibility paths with the provider-neutral cloud-drive alias. Document migration for existing configurations and keep the rclone manual QA coverage on the canonical module path.

Closes #1462

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)
