+++ src/datasource/skills/cloud-drive/index.ts
@@ -1,4 +1,10 @@
 export {
+	RcloneConnector,
+	type RcloneConnectorOptions,
+	type RcloneRunner,
+	type RcloneRunResult,
+} from "./rclone-connector.ts";
+export {
 	CLOUD_DRIVE_SKILL_DEFINITION,
 	CloudDriveSkill,
 	type CloudDriveSkillOptions,
