--- src/datasource/skills/gdrive/index.ts
@@ -1,8 +1,0 @@
-export { GDriveConnector, type GDriveConnectorOptions } from "./connector.ts";
-export {
-	RcloneConnector,
-	type RcloneConnectorOptions,
-	type RcloneRunner,
-	type RcloneRunResult,
-} from "./rclone-connector.ts";
-export { GDRIVE_SKILL_DEFINITION, GDriveSkill, type GDriveSkillOptions } from "./skill.ts";
